requirements.txt
transformers>=4.34.0
datasets>=2.14.5
torch>=2.0.1
pandas>=2.1.1
numpy>=1.24.3
scikit-learn>=1.3.1
gradio>=3.50.2
huggingface_hub>=0.17.3
python-dotenv>=1.0.0
pytest>=7.4.2
